requirements.txt
SpeechRecognition==3.10.0
pyautogui==0.9.54
pywinauto==0.6.8
psutil==5.9.6
pyttsx3==2.90
python-dotenv==1.0.0
pyaudio==0.2.14
openai==1.12.0
setuptools>=65.5.0
